Tests: Add tests
